kvm: put legacy driver code in kvm_legacy.c
