+++ libvmi/driver/kvm/CMakeLists.txt
@@ -1,6 +1,6 @@
 if (ENABLE_KVM_LEGACY)
     target_sources(vmi_shared PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/kvm.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/kvm_legacy.c
         ${CMAKE_CURRENT_SOURCE_DIR}/libvirt_wrapper.c)
 else ()
     # include libkvmi headers
